+++ app/src/main/java/ozarskiapps/booktracker/database/YearlyStatsDBService.kt
@@ -158,7 +158,7 @@ class YearlyStatsDBService(
         val bookStatus =
             BookStatus.valueOf(cursor.getString(cursor.getColumnIndexOrThrow(DatabaseConstants.BookTable.BOOK_STATUS_COLUMN)))
         val currentProgress =
-            cursor.getInt(cursor.getColumnIndexOrThrow(DatabaseConstants.BookTable.CURRENT_PROGRESS_COLUMN))
+            cursor.getFloat(cursor.getColumnIndexOrThrow(DatabaseConstants.BookTable.CURRENT_PROGRESS_COLUMN))
         val startDateMillis =
             cursor.getLong(cursor.getColumnIndexOrThrow(DatabaseConstants.BookTable.START_DATE_COLUMN))
         val endDateMillis =
